Had to split up preconditioned solve tests in epetra.
NOTE: It appears that when multiple sets are used in the parallel decomposition, this effectively hoses the epetra comm structure for later use (i.e. its static data gets out of whack somewhere.) this may cause issues in the future. Tpetra handles this fine by using the teuchos comm class.