Adding necessary and sufficient convergence test for Monte Carlo
Per Ji et all (2013), a necessary and sufficient convergence criteria
for the Neumann-Ulam algorithm is defined. That criteria has been
implemented along with other metrics from the paper. This adds
an Anasazi dependency because the metrics are eigenvalues of matrices.
Enable the criteria by setting the parameter:

"Compute Convergence Criteria"

to true in the input parameter list.